+++ Terraform/Azure/main.tf
@@ -1,53 +1,67 @@
 # First deploy for infrastructure, Unit Logics and networking.
 module "plan" {
   app         = var.plan
-  environment = var.environment
+  environment = local.environment
   group       = module.groups.apps
   source      = ".//modules//plan"
 }
 
 module "groups" {
   app         = var.group
-  environment = var.environment
+  environment = local.environment
   source      = ".//modules//groups"
 }
 
 module "sql" {
-  db          = var.db
-  sql         = var.sql
+  db          = local.db
+  sql         = local.sql
   source      = ".//modules//sql"
-  environment = var.environment
+  environment = local.environment
   group       = module.groups.storage
 }
 
+# ToDo: 20231114 - Deploy All secrets?:
+module "keys" {
+  active                          = local.active
+  created                         = local.created
+  location                        = local.location
+  assetname                       = local.assetname
+  environment                     = local.environment
+  group                           = module.groups.apps
+  source                          = ".//modules//keys"
+  key_vault_reference_identity_id = "39cee712-1020-496b-a490-57ab8f928149"
+  access_policies = [
+    {
+      tenant_id = "..."
+      object_id = "..."
+    },
+    # Agrega más políticas según sea necesario
+  ]
+}
+
 module "networking" {
-  apps        = var.apps
+  apps        = local.apps
   subnet      = var.subnet
   gateway     = var.gateway
-  network     = var.network
-  environment = var.environment
+  network     = local.network
+  created     = local.created
+  location    = local.location
+  assetname   = local.assetname
+  environment = local.environment
   group       = module.groups.networking
   source      = ".//modules//networking"
 }
 
-ToValidate: 20231123 - Deploy All secrets:
-module "keys" {
-  apps        = var.key
-  environment = var.environment
-  group       = module.groups.apps
-  source      = ".//modules//keys"
-}
-
 module "monitoring" {
-  apps        = var.apps
-  environment = var.environment
+  apps        = local.apps
+  environment = local.environment
   group       = module.groups.apps
   source      = ".//modules//monitoring"
 }
 
 module "storage" {
   app         = var.data
-  environment = var.environment
+  environment = local.environment
   group       = module.groups.storage
   source      = ".//modules//storage"
   function    = module.networking.subnet[0].id
@@ -57,8 +71,8 @@ module "dotnet" {
   app         = each.value
   for_each    = toset(var.net)
   plan        = module.plan.web
-  aspnetcore  = var.aspnetcore
-  environment = var.environment
+  aspnetcore  = local.aspnetcore
+  environment = local.environment
   group       = module.groups.apps
   source      = ".//modules//dotnet"
   key         = module.monitoring.key
@@ -67,52 +81,74 @@ module "dotnet" {
 }
 
 module "node" {
-  app         = each.value
-  for_each    = toset(var.node)
-  plan        = module.plan.web
-  aspnetcore  = var.aspnetcore
-  environment = var.environment
-  group       = module.groups.apps
-  source      = ".//modules//node"
-  key         = module.monitoring.key
-  connection  = module.monitoring.connection
-  public      = module.networking.gateway[0].id
+  connection       = module.monitoring.connection
+  override_special = local.override_special
+  key              = local.key
+  public           = module.networking.gateway[0].id
+  group            = module.groups.apps
+  tz               = local.tz
+  app              = each.value
+  active           = local.active
+  length           = local.length
+  created          = local.created
+  feature          = local.feature
+  location         = local.location
+  identity         = local.identity
+  assetname        = local.assetname
+  for_each         = toset(var.node)
+  plan             = module.plan.web
+  aspnetcore       = local.aspnetcore
+  appinsights      = local.appinsights
+  environment      = local.environment
+  source           = ".//modules//node"
 }
 
 module "functions" {
-  app         = each.value
-  aspnetcore  = var.aspnetcore
-  environment = var.environment
-  group       = module.groups.apps
-  for_each    = toset(var.function)
-  plan        = module.plan.function
-  key         = module.monitoring.key
-  storage     = module.storage.public
-  source      = ".//modules//functions"
-  connection  = module.monitoring.connection
-  private     = module.networking.subnet[0].id
-  pfunction   = module.networking.gateway[1].id
+  tz                  = local.tz
+  app                 = each.value
+  active              = local.active
+  length              = local.length
+  created             = local.created
+  feature             = local.feature
+  identity            = local.identity
+  location            = local.location
+  assetname           = local.assetname
+  aspnetcore          = local.aspnetcore
+  appinsights         = local.appinsights
+  environment         = local.environment
+  group               = module.groups.apps
+  for_each            = toset(var.function)
+  plan                = module.plan.function
+  key                 = module.monitoring.key
+  storage             = module.storage.private
+  source              = ".//modules//functions"
+  override_special    = local.override_special
+  connection          = module.monitoring.connection
+  private             = module.networking.subnet[0].id
+  public              = module.networking.gateway[0].id
+  pfunction           = module.networking.gateway[1].id
+  AzureWebJobsStorage = module.storage.AzureWebJobsStorage
 }
 
-ToValidate: 20231116 - Deploy Only Prod:
+# ToValidate: 20231116 - Deploy Only Prod:
 module "dns" {
   dns                 = var.dns
-  private             = var.pdns
+  private             = local.pdns
   source              = ".//modules//dns"
-  environment         = var.environment
+  environment         = local.environment
   group               = module.groups.networking
-  create_dns_resource = var.environment == "prod" ? true : false
+  create_dns_resource = local.environment == "prod" ? true : false
 }
 
 # Second deploy for compute.
 
 # Todo: 20231114 - Review Suscription Quotes and Deployments!!!
-#module "vms" {
-#  vm          = var.vm
-#  lvmss       = var.lvmss
-#  source      = ".//modules//vms"
-#  group       = module.groups.vms
-#  environment = var.environment
-#  set         = module.storage.set
-#  private     = module.networking.subnet[1].id
-#}
+# module "vms" {
+#   vm          = local.vm
+#   lvmss       = local.lvmss
+#   source      = ".//modules//vms"
+#   group       = module.groups.vms
+#   environment = local.environment
+#   set         = module.storage.set
+#   private     = module.networking.subnet[1].id
+# }